enhancing-dao-and-writing-interface-adapters
Enhancing the DAI for the make-team use-case by adding a saveTeam method, which in turn I implemented in our Persistence DAO (the implementation of the DAI). Also enhanced the structure of the csv file (reflected in the DAO code) to include a TEAM column, and solved minor problems surrounding that.